feat(ui): add new step 0 to select onchain/offchain type when creating a space
